Refactor: 모임 상세 조회 Repository -> Service로 매핑 변경
