+++ src/main/java/com/boardgo/domain/meeting/repository/response/MeetingDetailResponse.java
@@ -10,7 +10,6 @@ public record MeetingDetailResponse(
         Long meetingId,
         String userNickName,
         Double rating,
-        Long userWritingCount,
         LocalDateTime meetingDatetime,
         String likeStatus,
         String thumbnail,
