+++ src/main/java/com/boardgo/domain/meeting/service/MeetingCommandUseCase.java
@@ -8,4 +8,6 @@ public interface MeetingCommandUseCase {
     Long create(MeetingCreateRequest meetingCreateRequest, MultipartFile imageFile);
 
     void incrementShareCount(Long meetingId);
+
+    void incrementViewCount(Long meetingId);
 }
