+++ src/main/java/com/boardgo/domain/meeting/service/MeetingQueryServiceV1.java
@@ -2,15 +2,23 @@ package com.boardgo.domain.meeting.service;
 
 import com.boardgo.common.exception.CustomNoSuchElementException;
 import com.boardgo.common.utils.SecurityUtils;
+import com.boardgo.domain.boardgame.repository.BoardGameRepository;
+import com.boardgo.domain.boardgame.repository.response.BoardGameByMeetingIdResponse;
 import com.boardgo.domain.mapper.MeetingMapper;
 import com.boardgo.domain.meeting.controller.request.MeetingSearchRequest;
+import com.boardgo.domain.meeting.repository.MeetingLikeRepository;
 import com.boardgo.domain.meeting.repository.MeetingRepository;
+import com.boardgo.domain.meeting.repository.projection.MeetingDetailProjection;
 import com.boardgo.domain.meeting.repository.projection.MeetingSearchProjection;
 import com.boardgo.domain.meeting.repository.response.MeetingDetailResponse;
 import com.boardgo.domain.meeting.repository.response.MeetingSearchResponse;
+import com.boardgo.domain.review.repository.ReviewRepository;
+import com.boardgo.domain.user.repository.UserRepository;
+import com.boardgo.domain.user.repository.response.UserParticipantResponse;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -24,9 +32,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class MeetingQueryServiceV1 implements MeetingQueryUseCase {
     private final MeetingRepository meetingRepository;
     private final MeetingMapper meetingMapper;
+    private final UserRepository userRepository;
+    private final MeetingLikeRepository meetingLikeRepository;
+    private final ReviewRepository reviewRepository;
+    private final BoardGameRepository boardGameRepository;
 
     public Page<MeetingSearchResponse> search(MeetingSearchRequest searchRequest) {
-        // 페이지네이션 처리
         int size = getSize(searchRequest.size());
         int page = getPage(searchRequest.page());
         int offset = page * size;
@@ -56,13 +67,37 @@ public class MeetingQueryServiceV1 implements MeetingQueryUseCase {
     @Override
     public MeetingDetailResponse getDetailById(Long meetingId) {
         checkNullMeeting(meetingId);
-        MeetingDetailResponse result =
-                meetingRepository.findDetailById(
-                        meetingId, SecurityUtils.currentUserIdWithoutThrow());
+        Long userId = SecurityUtils.currentUserIdWithoutThrow();
 
-        meetingRepository.incrementViewCount(meetingId);
+        List<UserParticipantResponse> userParticipantResponseList =
+                userRepository.findByMeetingId(meetingId);
+        List<BoardGameByMeetingIdResponse> boardGameByMeetingIdResponseList =
+                boardGameRepository.findMeetingDetailByMeetingId(meetingId);
 
-        return result;
+        MeetingDetailProjection meetingDetailProjection =
+                meetingRepository.findDetailById(meetingId, userId);
+        Double rating = getRating(meetingDetailProjection);
+
+        return meetingMapper.toMeetingDetailResponse(
+                meetingDetailProjection,
+                userParticipantResponseList,
+                boardGameByMeetingIdResponseList,
+                meetingRepository.getCreateMeetingCount(meetingDetailProjection.userId()),
+                getLikeStatus(meetingId, userId),
+                rating);
+    }
+
+    private Double getRating(MeetingDetailProjection meetingDetailProjection) {
+        return Optional.ofNullable(
+                        reviewRepository.findRatingAvgByRevieweeId(
+                                meetingDetailProjection.userId()))
+                .orElse(0.0);
+    }
+
+    private String getLikeStatus(Long meetingId, Long userId) {
+        return Objects.nonNull(userId)
+                ? meetingLikeRepository.existsByUserIdAndMeetingId(userId, meetingId) ? "Y" : "N"
+                : "N";
     }
 
     private void checkNullMeeting(Long meetingId) {
