+++ src/test/java/com/boardgo/integration/meeting/controller/MeetingDocsTest.java
@@ -306,9 +306,6 @@ public class MeetingDocsTest extends RestDocsTestSupport {
                                         fieldWithPath("rating")
                                                 .type(JsonFieldType.NUMBER)
                                                 .description("모임 개설자의 평균 점수"),
-                                        fieldWithPath("userWritingCount")
-                                                .type(JsonFieldType.NUMBER)
-                                                .description("모임 개설자의 모임 개설 횟수"),
                                         fieldWithPath("title")
                                                 .type(JsonFieldType.STRING)
                                                 .description("모임 제목"),
